secures basic api routes

+++ server/api/users.js
@@ -5,26 +5,40 @@ const asyncHandler = require('express-async-handler')
 router.get(
   '/',
   asyncHandler(async (req, res, next) => {
-    console.log('here!')
-    const allUsers = await User.findAll({})
-    res.json(allUsers)
+    if (req.user !== undefined && req.user.dataValues.adminStatus === true) {
+      const allUsers = await User.findAll({})
+      res.json(allUsers)
+    }else{
+      res.send('You are not authorized!')
+    }
   })
 )
+
+
 //gets user by id
 router.get(
   '/:id',
   asyncHandler(async (req, res, next) => {
+    if (req.user !== undefined && req.user.dataValues.adminStatus === true) {
     const userId = req.params.id
     const user = await User.findAll({where: {id: userId}})
-    res.json(user[0])
+    res.json(user[0]) 
+  }else{
+      res.send('You are not authorized!')
+  }
   })
 )
+
 router.delete('/:id', async (req, res, next) => {
   try {
-    User.destroy({where: {id: req.params.id}})
-    res.sendStatus(202)
-  } catch (err) {
-    next(err)
-  }
+    if (req.user !== undefined && req.user.dataValues.adminStatus === true){
+      User.destroy({where: {id: req.params.id}})
+      res.sendStatus(202)
+    } else {
+      res.send('You are not authorized!')
+    }
+    } catch (err) {
+      next(err)
+    }
 })
 module.exports = router
